requirements.txt
alembic==1.7.7
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.22.3
pandas==1.4.2
pangres==4.1.1
pathspec==0.9.0
platformdirs==2.5.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
